requirements.txt
flask
flask-cors
torch
torchvision
pillow
huggingface_hub
numpy
